+++ StRoot/StHbtMaker/Cut/rotateToReactionPlaneEventCut.cxx
@@ -1,6 +1,6 @@
 /***************************************************************************
  *
- * $Id: rotateToReactionPlaneEventCut.cxx,v 1.1 2000/05/25 21:47:28 laue Exp $
+ * $Id: rotateToReactionPlaneEventCut.cxx,v 1.2 2001/04/02 16:15:48 jeromel Exp $
  *
  * Author: Mike Lisa, Ohio State, lisa@mps.ohio-state.edu
  ***************************************************************************
@@ -12,6 +12,9 @@
  ***************************************************************************
  *
  * $Log: rotateToReactionPlaneEventCut.cxx,v $
+ * Revision 1.2  2001/04/02 16:15:48  jeromel
+ * Type cast in sprintf() statement for bool conversion (insure++ issue)
+ *
  * Revision 1.1  2000/05/25 21:47:28  laue
  * new event cut which can be used to rotate an event around the z-axis
  *
@@ -69,7 +72,7 @@ StHbtString rotateToReactionPlaneEventCut::Report(){
   char Ctemp[100];
   sprintf(Ctemp,"rotateToReactionPlaneEventCut: ");
   Stemp += Ctemp;
-  sprintf(Ctemp,"\n Rotation :\t %d",mRotation);
+  sprintf(Ctemp,"\n Rotation :\t %d",(int) mRotation);
   Stemp += Ctemp;
   sprintf(Ctemp,"\n Multiplicity:\t %d-%d",mEventMult[0],mEventMult[1]);
   Stemp += Ctemp;
